backend/requirements.txt
Django>=4.2.7,<5.0
django-tenants>=3.5.0
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
uvicorn>=0.24.0
whitenoise>=6.6.0
sentry-sdk>=1.39.1
daphne>=4.0.0
Pillow==10.1.0
django-storages==1.14.2
django-redis==5.4.0
django-cacheops==6.1.0
django-health-check==3.17.0
djangorestframework-simplejwt==5.3.0 